refactor(commands): extract product tests to sibling file
